Acerto de Detalhes

+++ executa.php
@@ -10,12 +10,16 @@ function putFtpArray(array $arr,string $nome) {
 
 	file_put_contents($filename, json_encode($arr));
 
+	if (!file_exists($filename)) {
+		echo "Não foi possível achar o arquivo {$filename}\n";
+		exit;
+	}
+
 	$fd = ftp_connect($conf['host']);
 	$lg = ftp_login($fd, $conf['user'], $conf['passwd']);
 	ftp_pasv($fd, false	);
 	ftp_put($fd, "/conf/{$nome}", $filename, FTP_BINARY);
 	ftp_close($fd);
-	unlink($filename);
 }
 
 function getFtpArray(string $nome) : array {
@@ -28,6 +32,11 @@ function getFtpArray(string $nome) : array {
 	ftp_pasv($fd, false	);
 	ftp_get($fd, $filename, "/conf/{$nome}", FTP_BINARY);
 	ftp_close($fd);
+
+	if (!file_exists($filename)) {
+		echo "Não foi possível achar o arquivo {$filename}\n";
+		exit;
+	}
 
 	$var = json_decode(file_get_contents($filename),TRUE);
 	unlink($filename);
@@ -57,104 +66,120 @@ $tamanhos['20M'] = ['Hash'=>'66d8426057595b172e7a50be8ce65db7','Size'=>20000000]
 $tamanhos['25M'] = ['Hash'=>'8aa026b23a51940347335f5b22d0177b','Size'=>25000000];
 $tamanhos['30M'] = ['Hash'=>'1c0e814e642c5fd58a2ee3dcd8c9e807','Size'=>30000000];
 
-$solucoes ['Go-SergioCorreia_V2']  = ['exec'=>NULL,'language'=>'Go','creator'=>'Sérgio Correia','source'=>'https://github.com/OsProgramadores/op-desafios/blob/master/desafio-05/qrwteyrutiyoup/go'];
-$solucoes ['Go-SergioCorreia'] 	   = ['exec'=>'./Go-SergioCorreia_V2','language'=>'Go','creator'=>'Sérgio Correia','source'=>'https://github.com/OsProgramadores/op-desafios/blob/master/desafio-05/qrwteyrutiyoup/go'];
-$solucoes ['Go-SergioCorreia_V3']  = ['exec'=>NULL,'language'=>'Go','creator'=>'Sérgio Correia','source'=>'https://github.com/OsProgramadores/op-desafios/blob/master/desafio-05/qrwteyrutiyoup/go'];
-$solucoes ['C-SergioCorreia']	   = ['exec'=>'./C-SergioCorreia','language'=>'C','creator'=>'Sérgio Correia','source'=>'https://github.com/OsProgramadores/op-desafios/blob/master/desafio-05/qrwteyrutiyoup/c'];
-$solucoes ['Go-MarcoPaganini'] 	   = ['exec'=>'./Go-MarcoPaganini','language'=>'Go','creator'=>'Marco Paganini','source'=>'https://github.com/OsProgramadores/op-desafios/tree/master/desafio-05/marcopaganini'];
-$solucoes ['Php-Bcampos']	   = ['exec'=>'/usr/bin/php Php-Bcampos.php','language'=>'Php','creator'=>'Bernardino Campos','source'=>'https://github.com/BernardinoCampos/Desafio5/blob/master/Desafio5.php'];
-$solucoes ['Php-Bcampos-hhvm']	   = ['exec'=>NULL,'language'=>'Php','creator'=>'Bernardino Campos','source'=>'https://github.com/BernardinoCampos/Desafio5/blob/master/Desafio5.php'];
-$solucoes ['Php-Bcampos_V2']	   = ['exec'=>NULL,'language'=>'Php','creator'=>'Bernardino Campos','source'=>'https://github.com/BernardinoCampos/Desafio5/blob/master/Desafio5.php'];
-$solucoes ['Php-Mockba']	   = ['exec'=>'/usr/bin/php Php-Mockba.php','language'=>'Php','creator'=>'Mockba - The Borg','source'=>'https://github.com/BernardinoCampos/Desafio5/blob/master/Desafio5_V4.php'];
-$solucoes ['Java-MarcoAntonio']	   = ['exec'=>'/usr/bin/java -jar Java-MarcoAntonio.jar','language'=>'Java','creator'=>'Marco Antônio','source'=>'https://github.com/mrcrch/op-d05-java/tree/jsoniter2'];
-$solucoes ['Python-Demetrescu']	   = ['exec'=>'/usr/bin/python3 Python-Demetrescu.py','language'=>'Python','creator'=>'Roger Demetrescu','source'=>'https://github.com/rdemetrescu/OsProgramadores/tree/master/desafio-5'];
-$solucoes ['Python-LuizLima'] 	   = ['exec'=>'/usr/bin/python3 Python-LuizLima.py','language'=>'Python','creator'=>'Luiz Lima','source'=>''];
-$solucoes ['Python-CarlosAugusto'] = ['exec'=>NULL,'language'=>'Python','creator'=>'Carlos Augusto','source'=>'https://github.com/engaugusto/desafio_5_python/blob/master/main.py'];
-$solucoes ['Rust-AndreGarzia'] 	   = ['exec'=>'./Rust-AndreGarzia','language'=>'Rust','creator'=>'André Garzia','source'=>'https://bitbucket.org/andregarzia/desafio-5-rust'];
-$solucoes ['Rust-Leovano'] 	   = ['exec'=>'./Rust-Leovano','language'=>'Rust','creator'=>'Leo Silva Souza','source'=>'https://github.com/leovano/op-desafios/tree/master/desafio-05/leovano'];
-$solucoes ['C#-Washington'] 	   = ['exec'=>'dotnet C#-Washington.SRC/D5/bin/Release/netcoreapp2.0/D5.dll','language'=>'C#','creator'=>'Washington Ramos','source'=>''];
-$solucoes ['C#-Washington-V2'] 	   = ['exec'=>'dotnet C#-Washington.SRC/D5/bin/Release/netcoreapp2.0/D5-v2.dll','language'=>'C#','creator'=>'Washington Ramos','source'=>''];
+$solucoes ['Go-SergioCorreia_V2']	= ['exec'=>NULL,'env'=>[], 'language'=>'Go','creator'=>'Sérgio Correia','source'=>'https://github.com/OsProgramadores/op-desafios/blob/master/desafio-05/qrwteyrutiyoup/go'];
+#$solucoes ['Go-SergioCorreia']		= ['exec'=>'./Exec/Go-SergioCorreia_V2', 'env'=>['GOGC'=>'off'], 'language'=>'Go','creator'=>'Sérgio Correia','source'=>'https://github.com/OsProgramadores/op-desafios/blob/master/desafio-05/qrwteyrutiyoup/go'];
+$solucoes ['Go-SergioCorreia_V3']	= ['exec'=>NULL,'env'=>['GOGC'=>'off'], 'language'=>'Go','creator'=>'Sérgio Correia','source'=>'https://github.com/OsProgramadores/op-desafios/blob/master/desafio-05/qrwteyrutiyoup/go'];
+#$solucoes ['C-SergioCorreia']		= ['exec'=>'./Exec/C-SergioCorreia','env'=>[], 'language'=>'C','creator'=>'Sérgio Correia','source'=>'https://github.com/OsProgramadores/op-desafios/blob/master/desafio-05/qrwteyrutiyoup/c'];
+#$solucoes ['Go-MarcoPaganini']		= ['exec'=>'./Exec/Go-MarcoPaganini','env'=>['GOGC'=>'off'], 'language'=>'Go','creator'=>'Marco Paganini','source'=>'https://github.com/OsProgramadores/op-desafios/tree/master/desafio-05/marcopaganini'];
+#$solucoes ['Php-Bcampos']			= ['exec'=>'/usr/bin/php Exec/Php-Bcampos.php','env'=>[], 'language'=>'Php','creator'=>'Bernardino Campos','source'=>'https://github.com/BernardinoCampos/Desafio5/blob/master/Desafio5.php'];
+$solucoes ['Php-Bcampos-hhvm']		= ['exec'=>NULL,'env'=>[], 'language'=>'Php','creator'=>'Bernardino Campos','source'=>'https://github.com/BernardinoCampos/Desafio5/blob/master/Desafio5.php'];
+$solucoes ['Php-Bcampos_V2']		= ['exec'=>NULL,'env'=>[], 'language'=>'Php','creator'=>'Bernardino Campos','source'=>'https://github.com/BernardinoCampos/Desafio5/blob/master/Desafio5.php'];
+#$solucoes ['Php-Mockba']			= ['exec'=>'/usr/bin/php Exec/Php-Mockba.php','env'=>[], 'language'=>'Php','creator'=>'Mockba - The Borg','source'=>'https://github.com/BernardinoCampos/Desafio5/blob/master/Desafio5_V4.php'];
+#$solucoes ['Java-MarcoAntonio']	= ['exec'=>'/usr/bin/java -jar Exec/Java-MarcoAntonio.jar','env'=>['LANG'=>'pt_BR.UTF-8'], 'language'=>'Java','creator'=>'Marco Antônio','source'=>'https://github.com/mrcrch/op-d05-java/tree/jsoniter2'];
+#$solucoes ['Python-Demetrescu']	= ['exec'=>'/usr/bin/python3 Exec/Python-Demetrescu.py','env'=>[], 'language'=>'Python','creator'=>'Roger Demetrescu','source'=>'https://github.com/rdemetrescu/OsProgramadores/tree/master/desafio-5'];
+#$solucoes ['Python-LuizLima']		= ['exec'=>'/usr/bin/python3 Exec/Python-LuizLima.py','env'=>[], 'language'=>'Python','creator'=>'Luiz Lima','source'=>''];
+$solucoes ['Python-CarlosAugusto']	= ['exec'=>NULL,'env'=>[], 'language'=>'Python','creator'=>'Carlos Augusto','source'=>'https://github.com/engaugusto/desafio_5_python/blob/master/main.py'];
+#$solucoes ['Rust-AndreGarzia']		= ['exec'=>'./Exec/Rust-AndreGarzia','env'=>[], 'language'=>'Rust','creator'=>'André Garzia','source'=>'https://bitbucket.org/andregarzia/desafio-5-rust'];
+$solucoes ['Rust-Leovano2']			= ['exec'=>NULL,'env'=>[], 'language'=>'Rust','creator'=>'Leo Silva Souza','source'=>'https://github.com/leovano/op-desafios/tree/master/desafio-05/leovano'];
+#$solucoes ['Rust-Leovano']			= ['exec'=>'./Exec/Rust-Leovano2','env'=>[], 'language'=>'Rust','creator'=>'Leo Silva Souza','source'=>'https://github.com/leovano/op-desafios/tree/master/desafio-05/leovano'];
+#$solucoes ['C#-Washington']		= ['exec'=>'dotnet Sources/C#-Washington/D5/bin/Release/netcoreapp2.0/D5.dll','env'=>['LANG'=>'en_US.UTF-8'], 'language'=>'C#','creator'=>'Washington Ramos','source'=>''];
+#$solucoes ['C#-Washington-V2']		= ['exec'=>'dotnet C#-Washington.SRC/D5/bin/Release/netcoreapp2.0/D5-v2.dll','env'=>[], 'language'=>'C#','creator'=>'Washington Ramos','source'=>''];
+#$solucoes ['C#-RafaelPires']		= ['exec'=>'dotnet Sources/C#-RafaelPires/bin/Release/netcoreapp2.1/Desafio5.dll','env'=>[], 'language'=>'C#','creator'=>'Rafael Pires','source'=>'https://github.com/faelpires/op-desafios/tree/master/desafio-05/faelpires'];
 
-$numExecucoes = 10;
-#$numExecucoes = 5;
+#$numExecucoes = 10;
+$numExecucoes = 5;
+
 $limMem = 28;	// limite em GBytes
-
-foreach ($solucoes as $key => $value)
-	@unlink($key.".dat");
-
-$tempos = [];
-
-@system("rm -f ./HD/*");
-
-if (!file_exists('/sys/fs/cgroup/memory/OsProgramadores'))
-	@system ('/usr/bin/cgcreate -g memory:/OsProgramadores');
-
-@system ("echo \$(( {$limMem} * 1073741824 )) > /sys/fs/cgroup/memory/OsProgramadores/memory.limit_in_bytes");
-@system ("echo \$(( {$limMem} * 1073741824 )) > /sys/fs/cgroup/memory/OsProgramadores/memory.max_usage_in_bytes");
-@system ("echo 0 > /sys/fs/cgroup/memory/OsProgramadores/memory.swappiness");
-
-$descriptorspec = [
-   0 => ["pipe", "r"],
-   1 => ["file", "/var/tmp/resultado", "w"],
-   2 => ["file", "/var/tmp/lixo", "w"]
-];
-
-$env = ['GOGC'=>'off'];
-
-foreach ($tamanhos as $tam=>$v) {
-	$arquivo = "Funcionarios-{$tam}.json";
-
-	copy ("./Arquivos/{$arquivo}","./HD/{$arquivo}") or die('Não foi possível copiar o arquivo\n');
-	@system("sync");
-
-	for($aa=0;$aa<$numExecucoes;$aa++)
-		foreach ($solucoes as $key => $vector) {
-			if ($vector['exec']==NULL)
-				continue;
-			echo "Processando ($aa) $key [$tam] - ";
-			system('sync; echo 3 > /proc/sys/vm/drop_caches');
-			$cmd = "/usr/bin/cgexec -g memory:OsProgramadores {$vector['exec']} ./HD/{$arquivo}";
-
-			$tempo = microtime(True);
-			$process = proc_open($cmd,$descriptorspec,$pipes,getcwd(),$env);
-			fclose($pipes[0]);
-			$resultado = proc_close($process);
-			$tempo = microtime(True)-$tempo;
-
-			if ($resultado!=0) {
-				echo "Falhou - Tempo {$tempo}\n";
-				$tempos[$key][$tam][]=0;
-				continue;
-			}
-			echo "OK - Tempo {$tempo}\n";
-
-			$fd = popen("sort < /var/tmp/resultado | md5sum", 'r');
-            fscanf($fd,"%s %s",$hash,$lixo);
-            pclose($fd);
-
-			if ($hash!=$v['Hash']) {
-				echo "Hash de $key falhou em $tam\n";
-				$otal=0;
-			}
-
-			$tempos[$key][$tam][]=$tempo;
-		}
-	@unlink("./HD/{$arquivo}");
- }
 
 $resultados = getFtpArray('resultados.json');
 
-foreach($tempos as $solucao=>$a)
-	foreach($a as $tam=>$b)
-		$resultados[$solucao][$tam]=['media'=>array_sum($b)/Count($b), 'stdDev'=>stddev($b), 'stdError'=>(stddev($b)/sqrt(Count($b))), 'tempos'=>$b, 'Size'=>$tamanhos[$tam]['Size']];
+if (IsSet($solucoes)) {
+	foreach ($solucoes as $key => $value)
+		@unlink($key.".dat");
 
-foreach (getFtpArray('solucoes.json') as $key=>$sol) {
+	$tempos = [];
+
+	@system("rm -f ./HD/*");
+
+	if (!file_exists('/sys/fs/cgroup/memory/OsProgramadores'))
+		@system ('/usr/bin/cgcreate -g memory:/OsProgramadores');
+
+	@system ("echo \$(( {$limMem} * 1073741824 )) > /sys/fs/cgroup/memory/OsProgramadores/memory.limit_in_bytes");
+	@system ("echo \$(( {$limMem} * 1073741824 )) > /sys/fs/cgroup/memory/OsProgramadores/memory.max_usage_in_bytes");
+	@system ("echo 0 > /sys/fs/cgroup/memory/OsProgramadores/memory.swappiness");
+
+	$descriptorspec = [
+	   0 => ["pipe", "r"],
+	   1 => ["file", "/var/tmp/resultado", "w"],
+	   2 => ["file", "/var/tmp/lixo", "w"]
+	];
+
+	foreach ($tamanhos as $tam=>$v) {
+		$arquivo = "Funcionarios-{$tam}.json";
+
+		copy ("./Arquivos/{$arquivo}","./HD/{$arquivo}") or die('Não foi possível copiar o arquivo\n');
+		@system("sync");
+
+		for($aa=0;$aa<$numExecucoes;$aa++)
+			foreach ($solucoes as $key => $vector) {
+				if ($vector['exec']==NULL)
+					continue;
+
+				echo "Processando ($aa) $key [$tam] - ";
+				system('sync; echo 3 > /proc/sys/vm/drop_caches');
+				$cmd = "/usr/bin/cgexec -g memory:OsProgramadores {$vector['exec']} ./HD/{$arquivo}";
+
+				$tempo = microtime(True);
+				$process = proc_open($cmd,$descriptorspec,$pipes,getcwd(),$vector['env']);
+				fclose($pipes[0]);
+				$resultado = proc_close($process);
+				$tempo = microtime(True)-$tempo;
+
+				if ($resultado!=0) {
+					echo "Falhou [{$resultado}]- Tempo {$tempo}\n";
+					$tempos[$key][$tam][]=0;
+					continue;
+				}
+
+				echo "OK - Tempo {$tempo}\n";
+
+				$fd = popen("cat /var/tmp/resultado | sed '/^$/d' | sort | md5sum", 'r');
+	            fscanf($fd,"%s %s",$hash,$lixo);
+	            pclose($fd);
+
+				if ($hash!=$v['Hash']) {
+					echo "Hash de $key falhou em $tam - {$hash} {$v['Hash']} \n";
+					exit;
+					$otal=0;
+				}
+
+				$tempos[$key][$tam][]=$tempo;
+			}
+		@unlink("./HD/{$arquivo}");
+	 }
+	 foreach($tempos as $solucao=>$a)
+	 	foreach($a as $tam=>$b)
+	 		$resultados[$solucao][$tam]=['media'=>array_sum($b)/Count($b), 'stdDev'=>stddev($b), 'stdError'=>(stddev($b)/sqrt(Count($b))), 'tempos'=>$b, 'Size'=>$tamanhos[$tam]['Size']];
+}
+
+foreach (getFtpArray('solucoes.json') as $key=>$sol)
 	if (!isSet($solucoes[$key]))
 		$solucoes[$key] = $sol;
+
+foreach ($solucoes as $key=>$sol){
 	if ($sol['exec']===NULL) {
 		unset($solucoes[$key]);
 		unset($resultados[$key]);
 	}
+}
+
+$solucoes['LIMITE'] = ['exec'=>'','language'=>'LIMITE','creator'=>'Admin','source'=>''];
+$resultados['LIMITE'] = $resultados["Php-Bcampos"];
+
+foreach ($resultados['LIMITE'] as $key=>$value) {
+	$resultados['LIMITE'][$key]['media']*=1.2;
+	for ($ii=0; $ii<$numExecucoes; $ii++)
+		$resultados['LIMITE'][$key]['tempos'][$ii]*=1.2;
 }
 
 putFtpArray($solucoes,'solucoes.json');
